refactor: extensive optimization and cleanup of core systems
- Performance: Optimized leaderboard updates to only save when necessary. Reduced interval for payment cleanup.
- Structure: Moved UI constants and config schemas to dedicated files (`uiUtils.js`, `configPanelRegistry.js`).
- Cleanup: Extracted script event handling from `main.js` to `scriptEventReceive.js`. Removed redundant imports and unused variables.
- UI: Centralized pagination logic. Split "General" settings into sub-categories for better UX.
- Code Quality: Refactored `playerDataManager` for better modularity.
