+++ AddonExeBP/scripts/core/main.js
@@ -1,16 +1,21 @@
 import * as mc from '@minecraft/server';
-import { loadConfig, getConfig, updateConfig } from './configManager.js';
+import { loadConfig, getConfig } from './configManager.js';
 import { getSpawnConfig, loadEconomyConfig, loadKitsConfig, loadRanksConfig, loadShopConfig, loadSpawnConfig, loadXrayConfig } from './configurations.js';
 import * as dataManager from './dataManager.js';
 import * as rankManager from './rankManager.js';
-import * as playerDataManager from './playerDataManager.js';
-import { cleanupPlayerDataManager, clearExpiredPayments } from './playerDataManager.js';
+import {
+    getOrCreatePlayer,
+    setPlayerRank,
+    cleanupPlayerDataManager,
+    clearExpiredPayments,
+    loadNameIdMap,
+    initializeLeaderboard
+} from './playerDataManager.js';
 import { loadPunishments, clearExpiredPunishments, initializePunishmentManager } from './punishmentManager.js';
 import { loadReports, clearOldResolvedReports } from './reportManager.js';
 import { loadCooldowns, clearExpiredCooldowns } from './cooldownManager.js';
 import * as bountyManager from './bountyManager.js';
 import { errorLog, setLogLevel, infoLog } from './logger.js';
-import { startRestart } from './restartManager.js';
 import { initializeEventManager, cleanupEventManager } from './events/eventManager.js';
 import { cleanupTimers, setTrackedInterval } from './timerManager.js';
 import { initializeSpawnProtection } from '../modules/detections/spawnProtection.js';
@@ -25,7 +30,7 @@ import './mobDeathEvents.js';
  * @param {import('@minecraft/server').Player} player The player to check.
  */
 export function updatePlayerRank(player) {
-    const pData = playerDataManager.getOrCreatePlayer(player);
+    const pData = getOrCreatePlayer(player);
     if (!pData) { return; }
 
     const config = getConfig();
@@ -34,7 +39,7 @@ export function updatePlayerRank(player) {
     const newRank = rankManager.getPlayerRank(player, config);
 
     if (oldRankId !== newRank.id) {
-        playerDataManager.setPlayerRank(player.id, newRank.id, newRank.permissionLevel);
+        setPlayerRank(player.id, newRank.id, newRank.permissionLevel);
         infoLog(`[AddonExe] Player ${player.name}'s rank updated from ${oldRankId} to ${newRank.name}.`);
         player.sendMessage(`§aYour rank has been updated to ${newRank.name}.`);
     }
@@ -58,7 +63,7 @@ function reinitializeOnlinePlayers() {
     infoLog(`[AddonExe] Re-initializing state for ${mc.world.getAllPlayers().length} online players...`);
     for (const player of mc.world.getAllPlayers()) {
         // Ensure the player's data is loaded into the system
-        playerDataManager.getOrCreatePlayer(player);
+        getOrCreatePlayer(player);
         // Then, update their rank based on the loaded data and config
         updatePlayerRank(player);
     }
@@ -70,12 +75,12 @@ function reinitializeOnlinePlayers() {
  */
 function loadPersistentData() {
     infoLog('[AddonExe] Loading persistent data...');
-    playerDataManager.loadNameIdMap();
+    loadNameIdMap();
     loadPunishments();
     loadReports();
     loadCooldowns();
     bountyManager.loadBounties();
-    playerDataManager.initializeLeaderboard();
+    initializeLeaderboard();
 }
 
 /**
@@ -125,8 +130,8 @@ function checkConfiguration() {
  * Starts all recurring system timers.
  */
 function startSystemTimers() {
-    // Periodically clear expired payment confirmations
-    setTrackedInterval(clearExpiredPayments, 6000); // 5 minutes
+    // Periodically clear expired payment confirmations (every 60 seconds)
+    setTrackedInterval(clearExpiredPayments, 60 * 20);
     // Rank updates are now handled by events (e.g., !admin command)
     infoLog('[AddonExe] System timers started.');
 }
@@ -206,59 +211,12 @@ mc.system.runTimeout(async () => {
 }, 0);
 
 mc.system.afterEvents.scriptEventReceive.subscribe((event) => {
-    const { id, sourceEntity } = event;
+    const { id } = event;
 
     // Handle script unload event
     if (id === 'minecraft:script_unload') {
         cleanupAddon();
         return;
     }
-
-    const config = getConfig(); // Config should be loaded by the time this event fires for custom events.
-    if (!config) { return; }
-
-
-    switch (id) {
-        case 'exe:restart':
-            startRestart(sourceEntity);
-            break;
-
-        case 'exe:toggle_chat_log': {
-            const chatConfig = config.chat || { logToConsole: false };
-            const newValue = !chatConfig.logToConsole;
-            chatConfig.logToConsole = newValue;
-            updateConfig('chat', chatConfig);
-
-            const feedbackMessage = `§aChat-to-console has been ${newValue ? '§aenabled' : '§cdisabled'}§a.`;
-            if (sourceEntity && sourceEntity.sendMessage) {
-                sourceEntity.sendMessage(feedbackMessage);
-            }
-            // eslint-disable-next-line no-console
-            console.log(`[AddonExe] ${feedbackMessage}`);
-            break;
-        }
-
-        case 'exe:grant_admin_self': {
-            if (sourceEntity && sourceEntity.addTag) {
-                const adminRank = rankManager.getRankById('admin');
-                if (!adminRank) {
-                    errorLog('[AddonExe] Could not grant admin rank because the "admin" rank definition was not found.');
-                    sourceEntity.sendMessage('§cError: The admin rank is not configured.');
-                    return;
-                }
-
-                const adminTagCondition = adminRank.conditions.find(c => c.type === 'hasTag');
-                if (!adminTagCondition || !adminTagCondition.value) {
-                    errorLog('[AddonExe] Could not grant admin rank because it lacks a valid "hasTag" condition.');
-                    sourceEntity.sendMessage('§cError: The admin rank is not configured with a valid tag.');
-                    return;
-                }
-
-                sourceEntity.addTag(adminTagCondition.value);
-                sourceEntity.sendMessage('§aYou have been promoted to Admin.');
-                updateAllPlayerRanks();
-            }
-            break;
-        }
-    }
+    // Other script events are handled by handleScriptEventReceive in core/events/scriptEventReceive.js
 });
