+++ AddonExeBP/scripts/core/ui/panelHandlers.js
@@ -13,7 +13,7 @@ import { restartAnnouncer } from '../../modules/commands/announcement.js';
 import * as rulesManager from '../rulesManager.js';
 import * as helpfulLinksManager from '../helpfulLinksManager.js';
 import * as shopManager from '../shopManager.js';
-import { getKitsConfig, saveKitsConfig, getShopConfig, getSpawnConfig, saveSpawnConfig, getEconomyConfig, saveEconomyConfig, getXrayConfig, saveXrayConfig } from '../configurations.js';
+import { getKitsConfig, saveKitsConfig, getShopConfig, getEconomyConfig, saveEconomyConfig, getXrayConfig, saveXrayConfig } from '../configurations.js';
 import { items as allItems } from '../itemsConfig.js';
 import { createKit, deleteKit, getAllKits, updateKitSettings, renameKit } from '../kitAdminManager.js';
 import { addItemToKit, updateItemInKit } from '../kitItemsManager.js';
@@ -21,7 +21,7 @@ import * as shopAdminManager from '../shopAdminManager.js';
 import { initializeSpawnProtection } from '../../modules/detections/spawnProtection.js';
 import { showPanel } from '../uiManager.js';
 import { getVisiblePlayerActionItems, getMenuItems } from './panelBuilder.js';
-import { getVisibleConfigSystems } from './uiUtils.js';
+import { getVisibleConfigSystems, itemsPerPage, configHandlers, getPaginatedItems } from './uiUtils.js';
 import { panelDefinitions, configPanelSchema } from './panelRegistry.js';
 import { showConfirmationDialog } from './components.js';
 import { uiActionFunctions } from './actionRegistry.js';
@@ -31,38 +31,12 @@ import { spawnConfig as defaultSpawnConfig } from '../spawnConfig.js';
 import { economyConfig as defaultEconomyConfig } from '../economyConfig.js';
 import { xrayConfig as defaultXrayConfig } from '../xrayConfig.js';
 
-const itemsPerPage = 8;
 const allDefaultConfigs = {
     'main': defaultConfig,
     'spawn': defaultSpawnConfig,
     'economy': defaultEconomyConfig,
     'xray': defaultXrayConfig
 };
-
-const configHandlers = {
-    'main': {
-        get: getConfig,
-        save: (updates) => updateMultipleConfig(updates)
-    },
-    'spawn': {
-        get: getSpawnConfig,
-        save: (config) => saveSpawnConfig(config)
-    },
-    'economy': {
-        get: getEconomyConfig,
-        save: (config) => saveEconomyConfig(config)
-    },
-    'xray': {
-        get: getXrayConfig,
-        save: (config) => saveXrayConfig(config)
-    }
-};
-
-function getPaginatedItems(items, page) {
-    const startIndex = (page - 1) * itemsPerPage;
-    const endIndex = startIndex + itemsPerPage;
-    return items.slice(startIndex, endIndex);
-}
 
 export async function handleFormResponse(player, panelId, response, context) {
     const { selection, canceled, formValues } = response;
@@ -438,7 +412,9 @@ export async function handleFormResponse(player, panelId, response, context) {
     if (panelId === 'configResetPanel') {
         const page = context.page || 1;
         const resettableSystems = [
-            ...configPanelSchema.filter(c => c.id !== 'general').map(c => ({ id: c.id, title: c.title, icon: c.icon })),
+            ...configPanelSchema
+                .filter(c => !c.id.startsWith('general_')) // General settings are not individually resettable via this panel
+                .map(c => ({ id: c.id, title: c.title, icon: c.icon })),
             { id: 'kits', title: '§l§dKit System§r', icon: 'textures/ui/inventory_icon' },
             { id: 'shop', title: '§l§2Shop System§r', icon: 'textures/items/emerald' },
             { id: 'ranks', title: '§l§4Rank System§r', icon: 'textures/ui/permissions_member_star.png' }
