+++ AddonExeBP/scripts/core/ui/uiUtils.js
@@ -1,4 +1,55 @@
-import { configPanelSchema } from './panelRegistry.js';
+import { configPanelSchema } from './configPanelRegistry.js';
+import { getConfig, updateMultipleConfig } from '../configManager.js';
+import { getSpawnConfig, saveSpawnConfig, getEconomyConfig, saveEconomyConfig, getXrayConfig, saveXrayConfig } from '../configurations.js';
+
+export const itemsPerPage = 8;
+
+export const configHandlers = {
+    'main': {
+        get: getConfig,
+        save: (updates) => updateMultipleConfig(updates)
+    },
+    'spawn': {
+        get: getSpawnConfig,
+        save: (config) => saveSpawnConfig(config)
+    },
+    'economy': {
+        get: getEconomyConfig,
+        save: (config) => saveEconomyConfig(config)
+    },
+    'xray': {
+        get: getXrayConfig,
+        save: (config) => saveXrayConfig(config)
+    }
+};
+
+/**
+ * Helper to slice items for pagination.
+ * @param {Array} items
+ * @param {number} page
+ * @returns {Array}
+ */
+export function getPaginatedItems(items, page) {
+    const startIndex = (page - 1) * itemsPerPage;
+    const endIndex = startIndex + itemsPerPage;
+    return items.slice(startIndex, endIndex);
+}
+
+/**
+ * Helper to add pagination buttons to a form.
+ * @param {ActionFormData} form
+ * @param {number} page
+ * @param {number} totalItems
+ */
+export function addPaginationButtons(form, page, totalItems) {
+    const totalPages = Math.ceil(totalItems / itemsPerPage);
+    if (page > 1) {
+        form.button('§l§4< §1Previous');
+    }
+    if (page < totalPages) {
+        form.button('§l§1Next §4>');
+    }
+}
 
 /**
  * Generates a synchronized list of visible configuration systems for a player.
@@ -22,14 +73,24 @@ export function getVisibleConfigSystems(pData) {
         allSystems.push({ id: 'configResetPanel', title: '§l§cReset Settings§r', icon: 'textures/ui/wysiwyg_reset' });
     }
 
-    // Custom sorting: General first, Reset last, rest alphabetical
-    const generalSystem = allSystems.find(s => s.id === 'config_general');
+    // Custom sorting: Server Info first, Gameplay, System, Reset last, rest alphabetical
+    const serverInfo = allSystems.find(s => s.id === 'config_general_server');
+    const gameplay = allSystems.find(s => s.id === 'config_general_gameplay');
+    const system = allSystems.find(s => s.id === 'config_general_system');
     const resetSystem = allSystems.find(s => s.id === 'configResetPanel');
-    let otherSystems = allSystems.filter(s => s.id !== 'config_general' && s.id !== 'configResetPanel');
+
+    let otherSystems = allSystems.filter(s =>
+        s.id !== 'config_general_server' &&
+        s.id !== 'config_general_gameplay' &&
+        s.id !== 'config_general_system' &&
+        s.id !== 'configResetPanel'
+    );
     otherSystems.sort((a, b) => a.title.replace(/§./g, '').localeCompare(b.title.replace(/§./g, '')));
 
     const sortedSystems = [];
-    if (generalSystem) { sortedSystems.push(generalSystem); }
+    if (serverInfo) { sortedSystems.push(serverInfo); }
+    if (gameplay) { sortedSystems.push(gameplay); }
+    if (system) { sortedSystems.push(system); }
     sortedSystems.push(...otherSystems);
     if (resetSystem) { sortedSystems.push(resetSystem); }
